take the nominal search functionality and apply it to ocg search
